feat: #339 Created AfterExactlyStep and AfterAtLeastStep and tests.
(cherry picked from commit bc764e616f4a60574421c7173959642136d83e26)

+++ OcavaScenarioTest/src/test/java/com/ocadotechnology/scenario/TestEventWhenSteps.java
@@ -27,7 +27,7 @@ public class TestEventWhenSteps {
         this.simulationHolder = simulationHolder;
     }
 
-    public void scheduled(int time, String name) {
+    public void scheduled(double time, String name) {
         stepManager.addExecuteStep(() -> {
                 simulationHolder.getEventScheduler().doAt(
                         time,
